TT-440 Remove old document picker A/B test variant
Following the analysis of an A/B test for the document picker page, the B (new) variant has been chosen. This commit removes the old version and permanently moves to the new design.

Authors: @kgarwood @javindo @thomasalanlee
