+++ app/assets/javascripts/application.js.erb
@@ -19,7 +19,6 @@ window.GOVUK.willItWorkForMe.init();
 window.GOVUK.feedback.init();
 window.GOVUK.furtherInformation.init();
 window.GOVUK.interstitialQuestion.init();
-window.GOVUK.photoDocuments.init();
 window.GOVUK.otherDocuments.init();
 
 $(function () {
