+++ app/assets/javascripts/select_documents.js
@@ -1,71 +1,74 @@
-(function(global) {
-  "use strict"
-  var GOVUK = global.GOVUK || {};
-  var $ = global.jQuery;
+(function (global) {
+    "use strict";
+    var GOVUK = global.GOVUK || {};
+    var $ = global.jQuery;
 
-  var selectDocuments = {
-    markAllAsNo: function() {
-      // Mark all documents as 'No' when "I don't have documents" is selected
-      var $checkbox = $(this);
-      var checkboxValue = $checkbox.prop('checked');
-      var $noAnswers = selectDocuments.$form.find('input[type=radio][value=false]');
+    var selectDocuments = {
+        init: function () {
+            selectDocuments.$form = $('#validate-select-documents');
+            selectDocuments.$drivingLicenceDetails = $('#driving_licence_details');
+            var errorMessage = selectDocuments.$form.data('msg');
+            $.validator.addMethod('drivingLicenceDetailsValidation', function(value, element) {
+                return $('#select_documents_form_any_driving_licence_false').is(':checked') ||
+                    ($('#select_documents_form_any_driving_licence_true').is(':checked') && ($('#select_documents_form_driving_licence').is(':checked') || $('#select_documents_form_ni_driving_licence').is(':checked')))
+            }, $.validator.format(selectDocuments.$form.data('msg')));
 
-      if (checkboxValue === true) {
-        $noAnswers.trigger('click');
-        selectDocuments.speakWarningMessage();
-      } else {
-        selectDocuments.$noDocumentsMessage.empty();
-      }
-    },
-    unCheckNoDocuments: function() {
-      // Un check "I don't have documents" if the user selects a document
-      var $checkbox = selectDocuments.$form.find('.js-no-docs:checked');
-      $checkbox.prop('checked',false);
-    },
-    speakWarningMessage: function(){
-      // let screenreader users know their choices have changed by reading a notice
-      selectDocuments.$noDocumentsMessage.text(selectDocuments.$noDocumentsMessage.data('no-documents-message'));
-    },
-    init: function (){
-      selectDocuments.$form = $('#validate-documents');
-      selectDocuments.$noDocumentsMessage = $('#no-documents-message');
+            selectDocuments.validator = selectDocuments.$form.validate($.extend({}, GOVUK.validation.radiosValidation, {
+                rules: {
+                    'select_documents_form[any_driving_licence]': 'required',
+                    'select_documents_form[passport]': 'required',
+                    'select_documents_form[driving_licence]': 'drivingLicenceDetailsValidation',
+                    'select_documents_form[ni_driving_licence]': 'drivingLicenceDetailsValidation',
+                },
+                groups: {
+                    primary: 'select_documents_form[any_driving_licence] select_documents_form[passport] select_documents_form[driving_licence] select_documents_form[ni_driving_licence]'
+                },
+                messages: {
+                    'select_documents_form[any_driving_licence]': errorMessage,
+                    'select_documents_form[passport]': errorMessage,
+                    'select_documents_form[driving_licence]': errorMessage,
+                    'select_documents_form[ni_driving_licence]': errorMessage
+                }
+            }));
 
-      if (selectDocuments.$form.length === 1) {
-        selectDocuments.$form.find('.js-no-docs').on('click',selectDocuments.markAllAsNo);
-        selectDocuments.$form.find('input[type=radio][value=true]').on('click',selectDocuments.unCheckNoDocuments);
+            selectDocuments.setDrivingLicenceDetailsVisibility();
+            selectDocuments.$form.find('input[name="select_documents_form[any_driving_licence]"]')
+                .on('click', selectDocuments.setDrivingLicenceDetailsVisibility)
 
-        $.validator.addMethod('selectDocumentsValidation', function(value, element) {
-          var numberOfDocumentQuestions = selectDocuments.$form.find('.form-field').length;
-          var checkedElements = selectDocuments.$form.find('input[type=radio]').filter(':checked');
-          var allDocumentQuestionsAnswered = checkedElements.length === numberOfDocumentQuestions;
-          var hasAtLeastOneDocument = checkedElements.filter('[value=true]').length > 0;
-          return allDocumentQuestionsAnswered || hasAtLeastOneDocument;
-        }, $.validator.format(selectDocuments.$form.data('msg')));
+            // Changing driving license validation to false when there are validation errors
+            // on the driving license detail fields does not cause validation execution on the
+            // erroneous fields. This causes the main error message to remain visible even though
+            // there are no visible errors for the user.
+            //
+            // To work around this, we force the whole form validation on the above mentioned
+            // scenario.
+            selectDocuments.$form.find('#select_documents_form_any_driving_licence_false')
+                .on('change', function () {
+                    if (!GOVUK.selectDocuments.validator.valid()) {
+                        // the name of the api-function to validate the form is form :/
+                        // https://jqueryvalidation.org/Validator.form
+                        selectDocuments.validator.form()
+                    }
+                })
 
-        selectDocuments.$form.validate($.extend({}, GOVUK.validation.radiosValidation, {
-          rules: {
-            'select_documents_form[ni_driving_licence]': 'selectDocumentsValidation',
-            'select_documents_form[driving_licence]': 'selectDocumentsValidation',
-            'select_documents_form[passport]': 'selectDocumentsValidation',
-            'select_documents_form[non_uk_id_document]': 'selectDocumentsValidation',
-            'select_documents_form[no_documents]': 'selectDocumentsValidation'
-          },
-          groups: {
-            // driving_licence is the first element, error should focus this
-            driving_licence: 'select_documents_form[passport] select_documents_form[driving_licence] select_documents_form[ni_driving_licence] select_documents_form[non_uk_id_document] select_documents_form[no_documents]'
-          },
-          highlight: function(element, errorClass) {
-            selectDocuments.$form.children('.form-group:first').addClass('form-group-error');
-          },
-          unhighlight: function(element, errorClass) {
-            selectDocuments.$form.children('.form-group:first').removeClass('form-group-error');
-          }
-        }));
-      }
-    }
-  };
+        },
 
-  GOVUK.selectDocuments = selectDocuments;
+        setDrivingLicenceDetailsVisibility: function () {
+            if (selectDocuments.hasAnyDrivingLicence()) {
+                selectDocuments.$drivingLicenceDetails.removeClass('js-hidden');
+            } else {
+                selectDocuments.$drivingLicenceDetails.removeClass('form-group-error')
+                    .find('.selected').removeClass('selected').find('input').prop('checked', false);
+                selectDocuments.$drivingLicenceDetails.addClass('js-hidden');
+            }
+        },
+        hasAnyDrivingLicence: function () {
+            var input = $('input[name="select_documents_form[any_driving_licence]"]:checked');
+            return input.length === 1 && input.val() === 'true';
+        }
+    };
 
-  global.GOVUK = GOVUK;
+    GOVUK.selectDocuments = selectDocuments;
+
+    global.GOVUK = GOVUK;
 })(window);
