+++ app/views/select_documents/index.html.erb
@@ -1,61 +1,54 @@
-<%= page_title 'hub.select_documents.title' %>
+<%= page_title 'hub.select_documents.title_photo_documents' %>
 <% content_for :feedback_source, 'SELECT_DOCUMENTS_PAGE' %>
 
 <div id="no-documents-message" class="visually-hidden" aria-live="assertive" data-no-documents-message="<%= h t('hub.select_documents.no_documents_message') %>"></div>
 
 <div class="grid-row">
   <div class="column-two-thirds">
-    <h1 class="heading-large"><%= t 'hub.select_documents.title' %></h1>
-    <p><%= t 'hub.select_documents.explanation' %></p>
+    <h1 class="heading-large"><%= t 'hub.select_documents.title_photo_documents' %></h1>
+    <p><%= t 'hub.select_documents.sub_heading' %></p>
+    <div class="panel panel-border-narrow">
+      <%= t 'hub.select_documents.highlight' %>
+    </div>
+    <% if flash[:errors] %>
+        <div id="validation-error-message-no-js" class="validation-message form-group">
+          <p class="error"><%= flash[:errors] %></p>
+        </div>
+    <% end %>
 
-    <%= render 'shared/form-errors', errors: flash[:errors] %>
+    <%= form_for @form, url: select_documents_submit_path, html: {id: 'validate-select-documents', class: 'select-documents-form', novalidate: 'novalidate', 'data-msg': t('hub.select_documents.errors.no_selection')} do |f| %>
 
-    <%= form_for @form, url: select_documents_submit_path, html: { id: 'validate-documents', class: 'select-documents-form', novalidate: 'novalidate', 'data-msg': t('hub.select_documents.errors.no_selection')} do |f| %>
-        <%= content_tag :div, class: form_question_class do %>
-          <fieldset>
-            <legend><%= t 'hub.select_documents.legend'%></legend>
-            <div class="form-group form-field">
-              <fieldset class="inline">
-                <legend class="visually-hidden"><%= t 'hub.select_documents.question.passport'%></legend>
-                <span aria-hidden="true"><%= t 'hub.select_documents.question.passport'%></span>
+            <h2 class="heading-medium"><%= t 'hub.select_documents.valid_UK_driving_licence' %></h2>
+            <%= content_tag :div, class: form_question_class do %>
+              <fieldset>
+                <%= f.custom_radio_button :any_driving_licence, true, t('hub.select_documents.documents_option_yes') %>
+                <%= f.custom_radio_button :any_driving_licence, false, t('option.no') %>
+              </fieldset>
+            <% end %>
+            <%= content_tag :div, id: 'driving_licence_details', class: hidden_form_question_class do %>
+
+                <fieldset>
+                  <legend><span class="form-label"><%= t 'hub.select_documents.where_driving_licence_issued' %></span></legend>
+                    <%= f.custom_check_box :driving_licence, {class: 'form-group form-field'}, "true", "false", t('hub.select_documents.driving_licence_details.great_britain')  %>
+                    <%= f.custom_check_box :ni_driving_licence, {class: 'form-group form-field'}, "true", "false", t('hub.select_documents.driving_licence_details.northern_ireland')  %>
+                </fieldset>
+
+            <% end %>
+
+            <h2 class="heading-medium"><%= t 'hub.select_documents.uk_passport' %></h2>
+            <%= content_tag :div, class: form_question_class do %>
+              <fieldset>
                 <%= f.custom_radio_button :passport, true, t('hub.select_documents.documents_option_yes') %>
                 <%= f.custom_radio_button :passport, false, t('option.no') %>
               </fieldset>
-            </div>
-            <div class="form-group form-field">
-              <fieldset class="inline">
-                <legend class="visually-hidden"><%= t 'hub.select_documents.question.driving_licence'%></legend>
-                <span aria-hidden="true"><%= t 'hub.select_documents.question.driving_licence'%></span>
-                <%= f.custom_radio_button :driving_licence, true, t('hub.select_documents.documents_option_yes') %>
-                <%= f.custom_radio_button :driving_licence, false, t('option.no') %>
-              </fieldset>
-            </div>
-            <div class="form-group form-field">
-              <fieldset class="inline">
-                <legend class="visually-hidden"><%= t 'hub.select_documents.question.ni_driving_licence'%></legend>
-                <span aria-hidden="true"><%= t 'hub.select_documents.question.ni_driving_licence'%></span>
-                <%= f.custom_radio_button :ni_driving_licence, true, t('hub.select_documents.documents_option_yes') %>
-                <%= f.custom_radio_button :ni_driving_licence, false, t('option.no') %>
-              </fieldset>
-            </div>
-            <div class="form-group form-field">
-              <fieldset class="inline">
-                <legend class="visually-hidden"><%= t 'hub.select_documents.question.non_uk_id_document'%></legend>
-                <span aria-hidden="true"><%= t 'hub.select_documents.question.non_uk_id_document'%></span>
-                <%= f.custom_radio_button :non_uk_id_document, true, t('hub.select_documents.documents_option_yes') %>
-                <%= f.custom_radio_button :non_uk_id_document, false, t('option.no') %>
-              </fieldset>
-            </div>
+            <% end %>
 
-            <%= f.custom_check_box :no_documents, {class: 'js-no-docs'}, true, false, t('hub.select_documents.question.no_documents') %>
-
-          </fieldset>
-        <% end %>
 
         <div id="validation-error-message-js"></div>
 
         <div class="actions">
           <%= f.submit t('navigation.continue'), class: 'button', id: 'next-button' %>
+          <%= link_to t('hub.select_documents.neither_documents'), other_identity_documents_path %>
         </div>
     <% end %>
