+++ spec/features/user_visits_select_documents_page_test_spec.rb
@@ -1,16 +1,14 @@
 require 'feature_helper'
 require 'api_test_helper'
 
-RSpec.feature 'When B group user visits document selection page' do
+RSpec.feature 'When user visits document selection page' do
   before(:each) do
     set_session_and_session_cookies!
-    cookie_hash = create_cookie_hash.merge!(ab_test: CGI.escape({ 'split_questions_v2' => 'split_questions_v2_variant' }.to_json))
-    set_cookies!(cookie_hash)
     visit '/select-documents'
   end
 
   it 'includes the appropriate feedback source' do
-    expect_feedback_source_to_be(page, 'SELECT_DOCUMENTS_PAGE_PHOTO_DOCUMENTS', '/select-documents')
+    expect_feedback_source_to_be(page, 'SELECT_DOCUMENTS_PAGE', '/select-documents')
   end
 
   it 'should have a header about photo identity documents' do
@@ -22,17 +20,10 @@ RSpec.feature 'When B group user visits document selection page' do
     expect(page).to have_content('Eich dogfennau hunaniaeth gyda llun')
   end
 
-  it 'will report to piwki' do
-    piwik_request = {
-        '_cvar' => '{"6":["AB_TEST","split_questions_v2_variant"]}'
-    }
-    expect(a_request(:get, INTERNAL_PIWIK.url).with(query: hash_including(piwik_request))).to have_been_made.once
-  end
-
   it 'should go to select phone page when user has a valid GB licence or UK passport' do
-    choose 'photo_documents_form_any_driving_licence_true'
-    check 'photo_documents_form_driving_licence'
-    choose 'photo_documents_form_passport_true'
+    choose 'select_documents_form_any_driving_licence_true'
+    check 'select_documents_form_driving_licence'
+    choose 'select_documents_form_passport_true'
     click_button 'Continue'
     expect(page).to have_current_path(select_phone_path)
     expect(page.get_rack_session['selected_answers']).to eql('documents' => { 'passport' => true, 'driving_licence' => true, 'ni_driving_licence' => false })
@@ -46,8 +37,8 @@ RSpec.feature 'When B group user visits document selection page' do
     end
 
     it 'should go to other documents page if user does not have UK driving licence or UK passport' do
-      choose 'photo_documents_form_any_driving_licence_false'
-      choose 'photo_documents_form_passport_false'
+      choose 'select_documents_form_any_driving_licence_false'
+      choose 'select_documents_form_passport_false'
       click_button 'Continue'
       expect(page).to have_current_path(other_identity_documents_path)
       expect(page.get_rack_session['selected_answers']).to eql('documents' => { 'passport' => false, 'driving_licence' => false, 'ni_driving_licence' => false })
