+++ stub/ab_test.yml
@@ -31,9 +31,3 @@ experiments:
           percent: 0
         - name: 'no_logo_privacy'
           percent: 0
-  - split_questions_v2:
-      alternatives:
-        - name: 'control'
-          percent: 50
-        - name: 'variant'
-          percent: 50
